debug eau drone

+++ src/Robot.java
@@ -102,7 +102,7 @@ public abstract class Robot {
 
     /* Cree l'evenement pour que le robot se remplisse (en une fois) */
     public Evenement remplirEau(long date_absolue) {
-    	return new RemplirRobot(date_absolue+(int)this.getFullingTime(),this);
+    	return new RemplirRobot(date_absolue+(long)this.getFullingTime(),this);
     }
